........S. [ZBX-11951] Make trigger expressions with macros work again
This commit unbreak the functionality. Note that it will introduce some
performance hit due to the fact that we are potentially calling
substitute_simple_macro() two times in a code branch. My intention is
to have the fix in and iterate over this for re-working and improving
performances.
This patch ensure that all the macro are expanded before reaching
extract_expression_functionids(), that was the function failing due
to this.