........S. [ZBXNEXT-3661] First support of the REST API integration.
The implementation introduce a new configuration parameter in the
zabbix configuration file, "HistoryServiceURL".
This value will be used for enabling the usage of the new history
service integration. If the parameter is not set, the zabbix server will
continue to use the canonical zabbix database for storing history and
trends data.
The code flow for sending data to the history service, follow almost
the same logic of the existing one for committing data to the
zabbix database (particularly for trends).