........S. [ZBXNEXT-3661] Define stub functions for history service interfaces
Removed the series of #ifdef from dbcache.c and valuecache.c regarding the usage
of the compilation of the history service support. Instead, declared a series of stub
functions in history.c under a dedicated #ifdef that will be used if libCURL support
is not compiled in zabbix.
Also, moved the configuration check in src/server/server.c in the right place as
done for all the other features for being sure we do not use the history service
parameter without libCURL support.

Spotted by Andris Zeila.