........S. [ZBXNEXT-3661] Use dynamic buffers on JSON parsing
For safety, it's a good idea to allocate the buffers needed for processing JSON
data dynamically. In theory it should be enough to increase the static buffer
size, but we do not know in a future how the data management of our data
types will change.

Spotted by Andris Zeila.