........S. [ZBXNEXT-3877] Fix "null" field in history JSON output
In case the log->source field of a log value type was NULL, in the JSON sent
to the history service the value of the source field was the 'null' string.
Fix it by checking if the field is a NULL pointer and, if yes, set the value to
the empty string.
Decided this solution after speaking with Andris Zeila.