........S. [ZBXNEXT-3877] Fix modulo operation.
As the modulo operation between 0 and any divisor will always be 0, just
start the counter from 1 and use i - 1 for appending the trend to the vector
instead of checking for 0 before the sending condition at any iteration
(that condition will be true only once during the loop), thus avoiding the
double POST caused by this.
Also, use zbx_vector_ptr_clear for cleaning the vector and destroy it
only at the end.

Spotted by Andris Zeila.