........S. [ZBXNEXT-3877] merge the check for the history service URL with the types
Merge the HISTORY_SERVICE_URL series of checks in the function
zbx_history_check_type(). At the beginning I was reluctant to do this,
do to the fact that checking the type and the URL are two logically
and semantically separated operations.
However, after trying to implement this, the amount of boilerplate
removed far outweight the small logical problem.

Suggested by Andris Zeila.