........S. [ZBXNEXT-3877] Changed the types and url checks.
In r69561, the implemented series of checks was performed on two different
variables. One bitmask for data types and the history service URL. After
speaking again with Andris Zeila, merge the option checks in just the bitmask
variables by adding a flag for identifying if the history service usage is enabled
or not.