........S. [ZBXNEXT-4002] use elasticsearch scroll for retrieving values
Use of the scroll APIs allow us to process all the records needed regardless of
the elasticsearch intrinsic limitations.
The implementation keep track of the total number of records needed and stop
when the result of a scrolling query is empty. In this way we properly cover also
if a "count" parameter is not passed to the function.