........S. [ZBX-11426] do not delete events associated with a problem
Modify the housekeeper filters for the events for deleting only records that
are not associated with a record in the problem table (open or closed).
The remaining problems will be deleted after some time and also the lone
event records that will remain in the table.

Patch moved from the wrong development branch to here, contain also
style fixes by wiper.