........S. [ZBX-11426] added check for recovery events in problem record
The additional check ensure that we do not delete a recovery event
before the problem that it actually solve is deleted.
This can be problematic if the events housekeeping period is around
1 day.

To note that there an index need to be created for all the supported
databases, if not done, the performance impact could be potentially
huge on the housekeeper itself.

Solution proposed by Andris Zeila.