........S. [ZBXNEXT-4002] fixed trailing slash
Remove the trailing slash in the elasticsearch URL if used.
CURL does not seems to handle it for you.
Keep the full URL anyway as the main setting, because it's
possible that if in a future we implement other backends,
we want  to keep them integral as written in the config
file.